GrabIt/requirements.txt
Flask==2.0.1
PyMySQL==1.0.2
Werkzeug==2.0.1
python-dotenv==0.19.0
Flask-Session==0.4.0
Pillow==8.3.2